updated mocked_article
