--- src/constants/mockedArticle.ts
@@ -1,18 +1,0 @@
-export const MOCKED_ARTICLE = {
-  author: {
-    bio: null,
-    following: true,
-    image: 'https://static.productionready.io/images/smiley-cyrus.jpg',
-    username: 'mikeJax'
-  },
-  body: '<h2>Selección nacional</h2><p>Con la <a href="https://es.wikipedia.org/wiki/Selecci%C3%B3n_de_f%C3%BAtbol_de_B%C3%A9lgica" rel="noopener noreferrer" target="_blank">Selección de fútbol de Bélgica</a>, disputó 57 partidos internacionales y anotó solo 9 goles. Incluso participó con la selección belga, en 2 ediciones de la Copa Mundial. La primera participación de Van Moer en un mundial, fue en la edición de <a href="https://es.wikipedia.org/wiki/Copa_Mundial_de_F%C3%BAtbol_de_1970" rel="noopener noreferrer" target="_blank">México 1970</a>. donde su selección quedó eliminado, en la primera fase de la cita de <a href="https://es.wikipedia.org/wiki/M%C3%A9xico" rel="noopener noreferrer" target="_blank">México</a> y la segunda, fue en la edición de <a href="https://es.wikipedia.org/wiki/Copa_Mundial_de_F%C3%BAtbol_de_1982" rel="noopener noreferrer" target="_blank">España 1982</a>. donde su selección quedó eliminado, en la segunda fase de la cita de <a href="https://es.wikipedia.org/wiki/Espa%C3%B1a" rel="noopener noreferrer" target="_blank">España</a>. También participó en una sola edición de la <a href="https://es.wikipedia.org/wiki/Eurocopa" rel="noopener noreferrer" target="_blank">Eurocopa</a> y fue en la edición de <a href="https://es.wikipedia.org/wiki/Eurocopa_1980" rel="noopener noreferrer" target="_blank">Italia 1980</a>, cuando su selección perdió la Final de <a href="https://es.wikipedia.org/wiki/Roma" rel="noopener noreferrer" target="_blank">Roma</a>, ante su similar de <a href="https://es.wikipedia.org/wiki/Selecci%C3%B3n_de_f%C3%BAtbol_de_Alemania" rel="noopener noreferrer" target="_blank">Alemania</a>.</p>',
-  createdAt: '2021-08-26T14:51:05.900Z',
-  description:
-    'Wilfried Van Moer (Beveren, Bélgica, 1 de marzo de 1945-Lovaina, 24 de agosto de 2021)1​ fue un futbolista y entrenador belga, que jugaba de mediocampista y militó en diversos clubes de Bélgica. ',
-  favorited: false,
-  favoritesCount: 0,
-  slug: 'wilfried-van-moer-aznnm2',
-  tagList: [],
-  title: 'Wilfried Van Moer',
-  updatedAt: '2021-08-26T14:51:05.900Z'
-};
